@@ -14,23 +14,24 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  rxdart:
-  pdf: ^1.3.17
-  path_provider: any
-  sqfentity: ^1.2.1
-  sqfentity_gen: ^1.1.1
+  rxdart: ^0.27.3
+  pdf: ^3.7.3
+  path_provider: ^2.0.9
+  sqfentity: ^2.2.0+5
+  sqfentity_gen: ^2.2.0+10
   flutter:
     sdk: flutter
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
+  http: ^0.13.4
 
 dev_dependencies:
-  build_runner: ^1.6.5
-  build_verify: ^1.1.0
+  build_runner: ^2.1.8
+  build_verify: ^3.0.0
   flutter_test:
     sdk: flutter
 
