@@ -6,6 +6,8 @@ part of 'model.dart';
 // SqfEntityGenerator
 // **************************************************************************
 
+// ignore_for_file: non_constant_identifier_names
+
 //  These classes was generated by SqfEntity
 //  Copyright (c) 2019, All rights reserved. Use of this source code is governed by a
 //  Apache license that can be found in the LICENSE file.
@@ -37,7 +39,7 @@ class TableExpense extends SqfEntityTableBase {
     ];
     super.init();
   }
-  static SqfEntityTableBase _instance;
+  static SqfEntityTableBase? _instance;
   static SqfEntityTableBase get getInstance {
     return _instance = _instance ?? TableExpense();
   }
@@ -50,14 +52,14 @@ class SequenceIdentitySequence extends SqfEntitySequenceBase {
   SequenceIdentitySequence() {
     sequenceName = 'identity';
     maxValue =
-        9223372036854775807; /* optional. default is max int (9.223.372.036.854.775.807) */
+        9007199254740991; /* optional. default is max int (9.223.372.036.854.775.807) */
     cycle = false; /* optional. default is false; */
     minValue = 0; /* optional. default is 0 */
     incrementBy = 1; /* optional. default is 1 */
     startWith = 0; /* optional. default is 0 */
     super.init();
   }
-  static SequenceIdentitySequence _instance;
+  static SequenceIdentitySequence? _instance;
   static SequenceIdentitySequence get getInstance {
     return _instance = _instance ?? SequenceIdentitySequence();
   }
@@ -68,6 +70,10 @@ class SequenceIdentitySequence extends SqfEntitySequenceBase {
 class ExpenseModal extends SqfEntityModelProvider {
   ExpenseModal() {
     databaseName = expenseModel.databaseName;
+    password = expenseModel.password;
+    dbVersion = expenseModel.dbVersion;
+    preSaveAction = expenseModel.preSaveAction;
+    logFunction = expenseModel.logFunction;
     databaseTables = [
       TableExpense.getInstance,
     ];
@@ -78,9 +84,10 @@ class ExpenseModal extends SqfEntityModelProvider {
 
     bundledDatabasePath = expenseModel
         .bundledDatabasePath; //'assets/sample.db'; // This value is optional. When bundledDatabasePath is empty then EntityBase creats a new database when initializing the database
+    databasePath = expenseModel.databasePath;
   }
   Map<String, dynamic> getControllers() {
-    final controllers = Map<String, dynamic>();
+    final controllers = <String, dynamic>{};
 
     return controllers;
   }
@@ -89,233 +96,301 @@ class ExpenseModal extends SqfEntityModelProvider {
 
 // BEGIN ENTITIES
 // region Expense
-class Expense {
+class Expense extends TableBase {
   Expense(
       {this.id, this.amount, this.description, this.total, this.isDeleted}) {
-    setDefaultValues();
+    _setDefaultValues();
+    softDeleteActivated = true;
   }
   Expense.withFields(
       this.amount, this.description, this.total, this.isDeleted) {
-    setDefaultValues();
+    _setDefaultValues();
   }
   Expense.withId(
       this.id, this.amount, this.description, this.total, this.isDeleted) {
-    setDefaultValues();
+    _setDefaultValues();
   }
-  Expense.fromMap(Map<String, dynamic> o) {
-    id = o['id'] as int;
-    amount = o['amount'] as double;
-
-    description = o['description'] as String;
-
-    total = o['total'] as double;
-
-    isDeleted = o['isDeleted'] != null ? o['isDeleted'] == 1 : null;
-    isSaved = true;
+  // fromMap v2.0
+  Expense.fromMap(Map<String, dynamic> o, {bool setDefaultValues = true}) {
+    if (setDefaultValues) {
+      _setDefaultValues();
+    }
+    id = int.tryParse(o['id'].toString());
+    if (o['amount'] != null) {
+      amount = double.tryParse(o['amount'].toString());
+    }
+    if (o['description'] != null) {
+      description = o['description'].toString();
+    }
+    if (o['total'] != null) {
+      total = double.tryParse(o['total'].toString());
+    }
+    isDeleted = o['isDeleted'] != null
+        ? o['isDeleted'] == 1 || o['isDeleted'] == true
+        : null;
   }
-  // FIELDS
-  int id;
-  double amount;
-  String description;
-  double total;
-  bool isDeleted;
-  bool isSaved;
-  BoolResult saveResult;
-  // end FIELDS
+  // FIELDS (Expense)
+  int? id;
+  double? amount;
+  String? description;
+  double? total;
+  bool? isDeleted;
+
+  // end FIELDS (Expense)
 
   static const bool _softDeleteActivated = true;
-  ExpenseManager __mnExpense;
+  ExpenseManager? __mnExpense;
 
   ExpenseManager get _mnExpense {
     return __mnExpense = __mnExpense ?? ExpenseManager();
   }
 
-  // methods
-  Map<String, dynamic> toMap({bool forQuery = false, bool forJson = false}) {
-    final map = Map<String, dynamic>();
-    if (id != null) {
-      map['id'] = id;
-    }
-    if (amount != null) {
+  // METHODS
+  @override
+  Map<String, dynamic> toMap(
+      {bool forQuery = false, bool forJson = false, bool forView = false}) {
+    final map = <String, dynamic>{};
+    map['id'] = id;
+    if (amount != null || !forView) {
       map['amount'] = amount;
     }
-
-    if (description != null) {
+    if (description != null || !forView) {
       map['description'] = description;
     }
-
-    if (total != null) {
+    if (total != null || !forView) {
       map['total'] = total;
     }
-
     if (isDeleted != null) {
-      map['isDeleted'] = forQuery ? (isDeleted ? 1 : 0) : isDeleted;
+      map['isDeleted'] = forQuery ? (isDeleted! ? 1 : 0) : isDeleted;
     }
 
     return map;
   }
 
-  // methods
-  Future<Map<String, dynamic>> toMapWithChilds(
-      [bool forQuery = false, bool forJson = false]) async {
-    final map = Map<String, dynamic>();
-    if (id != null) {
-      map['id'] = id;
-    }
-    if (amount != null) {
+  @override
+  Future<Map<String, dynamic>> toMapWithChildren(
+      [bool forQuery = false,
+      bool forJson = false,
+      bool forView = false]) async {
+    final map = <String, dynamic>{};
+    map['id'] = id;
+    if (amount != null || !forView) {
       map['amount'] = amount;
     }
-
-    if (description != null) {
+    if (description != null || !forView) {
       map['description'] = description;
     }
-
-    if (total != null) {
+    if (total != null || !forView) {
       map['total'] = total;
     }
-
     if (isDeleted != null) {
-      map['isDeleted'] = forQuery ? (isDeleted ? 1 : 0) : isDeleted;
+      map['isDeleted'] = forQuery ? (isDeleted! ? 1 : 0) : isDeleted;
     }
 
     return map;
   }
 
-  /// This method always returns Json String
+  /// This method returns Json String [Expense]
+  @override
   String toJson() {
     return json.encode(toMap(forJson: true));
   }
 
-  /// This method always returns Json String
+  /// This method returns Json String [Expense]
+  @override
   Future<String> toJsonWithChilds() async {
-    return json.encode(await toMapWithChilds());
+    return json.encode(await toMapWithChildren(false, true));
   }
 
+  @override
   List<dynamic> toArgs() {
+    return [amount, description, total, isDeleted];
+  }
+
+  @override
+  List<dynamic> toArgsWithIds() {
     return [id, amount, description, total, isDeleted];
   }
 
-  static Future<List<Expense>> fromWebUrl(String url) async {
+  static Future<List<Expense>?> fromWebUrl(Uri uri,
+      {Map<String, String>? headers}) async {
     try {
-      final response = await http.get(url);
+      final response = await http.get(uri, headers: headers);
       return await fromJson(response.body);
     } catch (e) {
-      print(
+      debugPrint(
           'SQFENTITY ERROR Expense.fromWebUrl: ErrorMessage: ${e.toString()}');
       return null;
     }
   }
 
+  Future<http.Response> postUrl(Uri uri, {Map<String, String>? headers}) {
+    return http.post(uri, headers: headers, body: toJson());
+  }
+
   static Future<List<Expense>> fromJson(String jsonBody) async {
     final Iterable list = await json.decode(jsonBody) as Iterable;
-    var objList = List<Expense>();
+    var objList = <Expense>[];
     try {
       objList = list
           .map((expense) => Expense.fromMap(expense as Map<String, dynamic>))
           .toList();
     } catch (e) {
-      print('SQFENTITY ERROR Expense.fromJson: ErrorMessage: ${e.toString()}');
+      debugPrint(
+          'SQFENTITY ERROR Expense.fromJson: ErrorMessage: ${e.toString()}');
     }
     return objList;
   }
 
-  static Future<List<Expense>> fromObjectList(Future<List<dynamic>> o) async {
-    final expensesList = List<Expense>();
-    final data = await o;
-    for (int i = 0; i < data.length; i++) {
-      expensesList.add(Expense.fromMap(data[i] as Map<String, dynamic>));
-    }
-    return expensesList;
-  }
+  static Future<List<Expense>> fromMapList(List<dynamic> data,
+      {bool preload = false,
+      List<String>? preloadFields,
+      bool loadParents = false,
+      List<String>? loadedFields,
+      bool setDefaultValues = true}) async {
+    final List<Expense> objList = <Expense>[];
+    loadedFields = loadedFields ?? [];
+    for (final map in data) {
+      final obj = Expense.fromMap(map as Map<String, dynamic>,
+          setDefaultValues: setDefaultValues);
 
-  static List<Expense> fromMapList(List<Map<String, dynamic>> query) {
-    final List<Expense> expenses = List<Expense>();
-    for (Map map in query) {
-      expenses.add(Expense.fromMap(map as Map<String, dynamic>));
+      objList.add(obj);
     }
-    return expenses;
+    return objList;
   }
 
   /// returns Expense by ID if exist, otherwise returns null
-  /// <param name='id'>Primary Key Value</param>
-  /// <returns>returns Expense if exist, otherwise returns null
-  Future<Expense> getById(int id) async {
-    Expense expenseObj;
-    final data = await _mnExpense.getById(id);
-    if (data.length != 0) {
-      expenseObj = Expense.fromMap(data[0] as Map<String, dynamic>);
-    } else {
-      expenseObj = null;
+  /// Primary Keys: int? id
+  /// bool preload: if true, loads all related child objects (Set preload to true if you want to load all fields related to child or parent)
+  /// ex: getById(preload:true) -> Loads all related objects
+  /// List<String> preloadFields: specify the fields you want to preload (preload parameter's value should also be "true")
+  /// ex: getById(preload:true, preloadFields:['plField1','plField2'... etc])  -> Loads only certain fields what you specified
+  /// bool loadParents: if true, loads all parent objects until the object has no parent
+
+  /// <returns>returns [Expense] if exist, otherwise returns null
+  Future<Expense?> getById(int? id,
+      {bool preload = false,
+      List<String>? preloadFields,
+      bool loadParents = false,
+      List<String>? loadedFields}) async {
+    if (id == null) {
+      return null;
     }
-    return expenseObj;
+    Expense? obj;
+    final data = await _mnExpense.getById([id]);
+    if (data.length != 0) {
+      obj = Expense.fromMap(data[0] as Map<String, dynamic>);
+    } else {
+      obj = null;
+    }
+    return obj;
   }
 
-  /// Saves the object. If the id field is null, saves as a new record and returns new id, if id is not null then updates record
-
+  /// Saves the (Expense) object. If the id field is null, saves as a new record and returns new id, if id is not null then updates record
+  /// ignoreBatch = true as a default. Set ignoreBatch to false if you run more than one save() operation those are between batchStart and batchCommit
   /// <returns>Returns id
-  Future<int> save() async {
-    if (id == null || id == 0 || !isSaved) {
-      id = await _mnExpense.insert(this);
-      isSaved = true;
+  @override
+  Future<int?> save({bool ignoreBatch = true}) async {
+    if (id == null || id == 0) {
+      id = await _mnExpense.insert(this, ignoreBatch);
     } else {
-      id = await _upsert();
+      await _mnExpense.update(this);
     }
+
+    return id;
+  }
+
+  /// Saves the (Expense) object. If the id field is null, saves as a new record and returns new id, if id is not null then updates record
+  /// ignoreBatch = true as a default. Set ignoreBatch to false if you run more than one save() operation those are between batchStart and batchCommit
+  /// <returns>Returns id
+  @override
+  Future<int?> saveOrThrow({bool ignoreBatch = true}) async {
+    if (id == null || id == 0) {
+      id = await _mnExpense.insertOrThrow(this, ignoreBatch);
+
+      isInsert = true;
+    } else {
+      // id= await _upsert(); // removed in sqfentity_gen 1.3.0+6
+      await _mnExpense.updateOrThrow(this);
+    }
+
     return id;
   }
 
   /// saveAs Expense. Returns a new Primary Key value of Expense
 
   /// <returns>Returns a new Primary Key value of Expense
-  Future<int> saveAs() async {
-    isSaved = false;
+  @override
+  Future<int?> saveAs({bool ignoreBatch = true}) async {
     id = null;
-    return save();
+
+    return save(ignoreBatch: ignoreBatch);
   }
 
-  /// saveAll method saves the sent List<Expense> as a batch in one transaction
+  /// saveAll method saves the sent List<Expense> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  Future<List<BoolResult>> saveAll(List<Expense> expenses) async {
-    final results = _mnExpense.saveAll(
-        'INSERT OR REPLACE INTO expense (id,  amount, description, total,isDeleted)  VALUES (?,?,?,?,?)',
-        expenses);
-    return results;
+  static Future<List<dynamic>> saveAll(List<Expense> expenses) async {
+    List<dynamic>? result = [];
+    // If there is no open transaction, start one
+    final isStartedBatch = await ExpenseModal().batchStart();
+    for (final obj in expenses) {
+      await obj.save(ignoreBatch: false);
+    }
+    if (!isStartedBatch) {
+      result = await ExpenseModal().batchCommit();
+      for (int i = 0; i < expenses.length; i++) {
+        if (expenses[i].id == null) {
+          expenses[i].id = result![i] as int;
+        }
+      }
+    }
+    return result!;
   }
 
   /// Updates if the record exists, otherwise adds a new row
-
   /// <returns>Returns id
-  Future<int> _upsert() async {
+  @override
+  Future<int?> upsert({bool ignoreBatch = true}) async {
     try {
-      id = await _mnExpense.rawInsert(
-          'INSERT OR REPLACE INTO expense (id,  amount, description, total,isDeleted)  VALUES (?,?,?,?,?)',
-          [id, amount, description, total, isDeleted]);
-      saveResult = BoolResult(
-          success: true, successMessage: 'Expense id=$id updated successfuly');
+      final result = await _mnExpense.rawInsert(
+          'INSERT OR REPLACE INTO expense (id, amount, description, total,isDeleted)  VALUES (?,?,?,?,?)',
+          [id, amount, description, total, isDeleted],
+          ignoreBatch);
+      if (result! > 0) {
+        saveResult = BoolResult(
+            success: true,
+            successMessage: 'Expense id=$id updated successfully');
+      } else {
+        saveResult = BoolResult(
+            success: false, errorMessage: 'Expense id=$id did not update');
+      }
       return id;
     } catch (e) {
       saveResult = BoolResult(
           success: false,
           errorMessage: 'Expense Save failed. Error: ${e.toString()}');
-      return 0;
+      return null;
     }
   }
 
-  /// inserts or replaces the sent List<Todo> as a batch in one transaction.
+  /// inserts or replaces the sent List<<Expense>> as a bulk in one transaction.
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
-  /// Returns a <List<BoolResult>>
-  Future<List<BoolResult>> upsertAll(List<Expense> expenses) async {
+  /// Returns a BoolCommitResult
+  @override
+  Future<BoolCommitResult> upsertAll(List<Expense> expenses) async {
     final results = await _mnExpense.rawInsertAll(
-        'INSERT OR REPLACE INTO expense (id,  amount, description, total,isDeleted)  VALUES (?,?,?,?,?)',
+        'INSERT OR REPLACE INTO expense (id, amount, description, total,isDeleted)  VALUES (?,?,?,?,?)',
         expenses);
     return results;
   }
 
   /// Deletes Expense
 
-  /// <returns>BoolResult res.success=Deleted, not res.success=Can not deleted
+  /// <returns>BoolResult res.success= true (Deleted), false (Could not be deleted)
+  @override
   Future<BoolResult> delete([bool hardDelete = false]) async {
-    print('SQFENTITIY: delete Expense invoked (id=$id)');
-    if (!_softDeleteActivated || hardDelete || isDeleted) {
+    debugPrint('SQFENTITIY: delete Expense invoked (id=$id)');
+    if (!_softDeleteActivated || hardDelete || isDeleted!) {
       return _mnExpense
           .delete(QueryParams(whereString: 'id=?', whereArguments: [id]));
     } else {
@@ -325,11 +400,12 @@ class Expense {
     }
   }
 
-  /// Recover Expense>
+  /// Recover Expense
 
   /// <returns>BoolResult res.success=Recovered, not res.success=Can not recovered
+  @override
   Future<BoolResult> recover([bool recoverChilds = true]) async {
-    print('SQFENTITIY: recover Expense invoked (id=$id)');
+    debugPrint('SQFENTITIY: recover Expense invoked (id=$id)');
     {
       return _mnExpense.updateBatch(
           QueryParams(whereString: 'id=?', whereArguments: [id]),
@@ -337,489 +413,346 @@ class Expense {
     }
   }
 
-  //private ExpenseFilterBuilder _Select;
+  @override
   ExpenseFilterBuilder select(
-      {List<String> columnsToSelect, bool getIsDeleted}) {
-    return ExpenseFilterBuilder(this)
-      .._getIsDeleted = getIsDeleted == true
+      {List<String>? columnsToSelect, bool? getIsDeleted}) {
+    return ExpenseFilterBuilder(this, getIsDeleted)
       ..qparams.selectColumns = columnsToSelect;
   }
 
+  @override
   ExpenseFilterBuilder distinct(
-      {List<String> columnsToSelect, bool getIsDeleted}) {
-    return ExpenseFilterBuilder(this)
-      .._getIsDeleted = getIsDeleted == true
+      {List<String>? columnsToSelect, bool? getIsDeleted}) {
+    return ExpenseFilterBuilder(this, getIsDeleted)
       ..qparams.selectColumns = columnsToSelect
       ..qparams.distinct = true;
   }
 
-  void setDefaultValues() {
-    isSaved = false;
+  void _setDefaultValues() {
     isDeleted = isDeleted ?? false;
   }
-  //end methods
+
+  @override
+  void rollbackPk() {
+    if (isInsert == true) {
+      id = null;
+    }
+  }
+
+  // END METHODS
+  // BEGIN CUSTOM CODE
+  /*
+      you can define customCode property of your SqfEntityTable constant. For example:
+      const tablePerson = SqfEntityTable(
+      tableName: 'person',
+      primaryKeyName: 'id',
+      primaryKeyType: PrimaryKeyType.integer_auto_incremental,
+      fields: [
+        SqfEntityField('firstName', DbType.text),
+        SqfEntityField('lastName', DbType.text),
+      ],
+      customCode: '''
+       String fullName()
+       { 
+         return '$firstName $lastName';
+       }
+      ''');
+     */
+  // END CUSTOM CODE
 }
 // endregion expense
 
 // region ExpenseField
-class ExpenseField extends SearchCriteria {
-  ExpenseField(this.expenseFB) {
-    param = DbParameter();
-  }
-  DbParameter param;
-  String _waitingNot = '';
-  ExpenseFilterBuilder expenseFB;
+class ExpenseField extends FilterBase {
+  ExpenseField(ExpenseFilterBuilder expenseFB) : super(expenseFB);
 
-  ExpenseField get not {
-    _waitingNot = ' NOT ';
-    return this;
+  @override
+  ExpenseFilterBuilder equals(dynamic pValue) {
+    return super.equals(pValue) as ExpenseFilterBuilder;
   }
 
-  ExpenseFilterBuilder equals(var pValue) {
-    param.expression = '=';
-    expenseFB._addedBlocks = _waitingNot == ''
-        ? setCriteria(pValue, expenseFB.parameters, param, SqlSyntax.EQuals,
-            expenseFB._addedBlocks)
-        : setCriteria(pValue, expenseFB.parameters, param, SqlSyntax.NotEQuals,
-            expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+  @override
+  ExpenseFilterBuilder equalsOrNull(dynamic pValue) {
+    return super.equalsOrNull(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder isNull() {
-    expenseFB._addedBlocks = setCriteria(
-        0,
-        expenseFB.parameters,
-        param,
-        SqlSyntax.IsNULL.replaceAll(SqlSyntax.notKeyword, _waitingNot),
-        expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+    return super.isNull() as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder contains(dynamic pValue) {
-    if (pValue != null) {
-      expenseFB._addedBlocks = setCriteria(
-          '%${pValue.toString()}%',
-          expenseFB.parameters,
-          param,
-          SqlSyntax.Contains.replaceAll(SqlSyntax.notKeyword, _waitingNot),
-          expenseFB._addedBlocks);
-      _waitingNot = '';
-      expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-          expenseFB._addedBlocks.retVal;
-    }
-    return expenseFB;
+    return super.contains(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder startsWith(dynamic pValue) {
-    if (pValue != null) {
-      expenseFB._addedBlocks = setCriteria(
-          '${pValue.toString()}%',
-          expenseFB.parameters,
-          param,
-          SqlSyntax.Contains.replaceAll(SqlSyntax.notKeyword, _waitingNot),
-          expenseFB._addedBlocks);
-      _waitingNot = '';
-      expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-          expenseFB._addedBlocks.retVal;
-      expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-          expenseFB._addedBlocks.retVal;
-    }
-    return expenseFB;
+    return super.startsWith(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder endsWith(dynamic pValue) {
-    if (pValue != null) {
-      expenseFB._addedBlocks = setCriteria(
-          '%${pValue.toString()}',
-          expenseFB.parameters,
-          param,
-          SqlSyntax.Contains.replaceAll(SqlSyntax.notKeyword, _waitingNot),
-          expenseFB._addedBlocks);
-      _waitingNot = '';
-      expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-          expenseFB._addedBlocks.retVal;
-    }
-    return expenseFB;
+    return super.endsWith(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder between(dynamic pFirst, dynamic pLast) {
-    if (pFirst != null && pLast != null) {
-      expenseFB._addedBlocks = setCriteria(
-          pFirst,
-          expenseFB.parameters,
-          param,
-          SqlSyntax.Between.replaceAll(SqlSyntax.notKeyword, _waitingNot),
-          expenseFB._addedBlocks,
-          pLast);
-    } else if (pFirst != null) {
-      if (_waitingNot != '') {
-        expenseFB._addedBlocks = setCriteria(pFirst, expenseFB.parameters,
-            param, SqlSyntax.LessThan, expenseFB._addedBlocks);
-      } else {
-        expenseFB._addedBlocks = setCriteria(pFirst, expenseFB.parameters,
-            param, SqlSyntax.GreaterThanOrEquals, expenseFB._addedBlocks);
-      }
-    } else if (pLast != null) {
-      if (_waitingNot != '') {
-        expenseFB._addedBlocks = setCriteria(pLast, expenseFB.parameters, param,
-            SqlSyntax.GreaterThan, expenseFB._addedBlocks);
-      } else {
-        expenseFB._addedBlocks = setCriteria(pLast, expenseFB.parameters, param,
-            SqlSyntax.LessThanOrEquals, expenseFB._addedBlocks);
-      }
-    }
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+    return super.between(pFirst, pLast) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder greaterThan(dynamic pValue) {
-    param.expression = '>';
-    expenseFB._addedBlocks = _waitingNot == ''
-        ? setCriteria(pValue, expenseFB.parameters, param,
-            SqlSyntax.GreaterThan, expenseFB._addedBlocks)
-        : setCriteria(pValue, expenseFB.parameters, param,
-            SqlSyntax.LessThanOrEquals, expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+    return super.greaterThan(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder lessThan(dynamic pValue) {
-    param.expression = '<';
-    expenseFB._addedBlocks = _waitingNot == ''
-        ? setCriteria(pValue, expenseFB.parameters, param, SqlSyntax.LessThan,
-            expenseFB._addedBlocks)
-        : setCriteria(pValue, expenseFB.parameters, param,
-            SqlSyntax.GreaterThanOrEquals, expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+    return super.lessThan(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder greaterThanOrEquals(dynamic pValue) {
-    param.expression = '>=';
-    expenseFB._addedBlocks = _waitingNot == ''
-        ? setCriteria(pValue, expenseFB.parameters, param,
-            SqlSyntax.GreaterThanOrEquals, expenseFB._addedBlocks)
-        : setCriteria(pValue, expenseFB.parameters, param, SqlSyntax.LessThan,
-            expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+    return super.greaterThanOrEquals(pValue) as ExpenseFilterBuilder;
   }
 
+  @override
   ExpenseFilterBuilder lessThanOrEquals(dynamic pValue) {
-    param.expression = '<=';
-    expenseFB._addedBlocks = _waitingNot == ''
-        ? setCriteria(pValue, expenseFB.parameters, param,
-            SqlSyntax.LessThanOrEquals, expenseFB._addedBlocks)
-        : setCriteria(pValue, expenseFB.parameters, param,
-            SqlSyntax.GreaterThan, expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+    return super.lessThanOrEquals(pValue) as ExpenseFilterBuilder;
   }
 
-  ExpenseFilterBuilder inValues(var pValue) {
-    expenseFB._addedBlocks = setCriteria(
-        pValue,
-        expenseFB.parameters,
-        param,
-        SqlSyntax.IN.replaceAll(SqlSyntax.notKeyword, _waitingNot),
-        expenseFB._addedBlocks);
-    _waitingNot = '';
-    expenseFB._addedBlocks.needEndBlock[expenseFB._blockIndex] =
-        expenseFB._addedBlocks.retVal;
-    return expenseFB;
+  @override
+  ExpenseFilterBuilder inValues(dynamic pValue) {
+    return super.inValues(pValue) as ExpenseFilterBuilder;
+  }
+
+  @override
+  ExpenseField get not {
+    return super.not as ExpenseField;
   }
 }
 // endregion ExpenseField
 
 // region ExpenseFilterBuilder
-class ExpenseFilterBuilder extends SearchCriteria {
-  ExpenseFilterBuilder(Expense obj) {
-    whereString = '';
-    qparams = QueryParams();
-    parameters = List<DbParameter>();
-    orderByList = List<String>();
-    groupByList = List<String>();
-    _addedBlocks = AddedBlocks(List<bool>(), List<bool>());
-    _addedBlocks.needEndBlock.add(false);
-    _addedBlocks.waitingStartBlock.add(false);
-    _pagesize = 0;
-    _page = 0;
-    _obj = obj;
+class ExpenseFilterBuilder extends ConjunctionBase {
+  ExpenseFilterBuilder(Expense obj, bool? getIsDeleted)
+      : super(obj, getIsDeleted) {
+    _mnExpense = obj._mnExpense;
+    _softDeleteActivated = obj.softDeleteActivated;
   }
-  AddedBlocks _addedBlocks;
-  int _blockIndex = 0;
-  List<DbParameter> parameters;
-  List<String> orderByList;
-  Expense _obj;
-  QueryParams qparams;
-  int _pagesize;
-  int _page;
 
+  bool _softDeleteActivated = false;
+  ExpenseManager? _mnExpense;
+
+  /// put the sql keyword 'AND'
+  @override
   ExpenseFilterBuilder get and {
-    if (parameters.isNotEmpty) {
-      parameters[parameters.length - 1].wOperator = ' AND ';
-    }
+    super.and;
     return this;
   }
 
+  /// put the sql keyword 'OR'
+  @override
   ExpenseFilterBuilder get or {
-    if (parameters.isNotEmpty) {
-      parameters[parameters.length - 1].wOperator = ' OR ';
-    }
+    super.or;
     return this;
   }
 
+  /// open parentheses
+  @override
   ExpenseFilterBuilder get startBlock {
-    _addedBlocks.waitingStartBlock.add(true);
-    _addedBlocks.needEndBlock.add(false);
-    _blockIndex++;
-    if (_blockIndex > 1) _addedBlocks.needEndBlock[_blockIndex - 1] = true;
+    super.startBlock;
     return this;
   }
 
-  ExpenseFilterBuilder where(String whereCriteria) {
-    if (whereCriteria != null && whereCriteria != '') {
-      final DbParameter param = DbParameter();
-      _addedBlocks =
-          setCriteria(0, parameters, param, '($whereCriteria)', _addedBlocks);
-      _addedBlocks.needEndBlock[_blockIndex] = _addedBlocks.retVal;
-    }
+  /// String whereCriteria, write raw query without 'where' keyword. Like this: 'field1 like 'test%' and field2 = 3'
+  @override
+  ExpenseFilterBuilder where(String? whereCriteria, {dynamic parameterValue}) {
+    super.where(whereCriteria, parameterValue: parameterValue);
     return this;
   }
 
+  /// page = page number,
+  /// pagesize = row(s) per page
+  @override
   ExpenseFilterBuilder page(int page, int pagesize) {
-    if (page > 0) _page = page;
-    if (pagesize > 0) _pagesize = pagesize;
+    super.page(page, pagesize);
     return this;
   }
 
+  /// int count = LIMIT
+  @override
   ExpenseFilterBuilder top(int count) {
-    if (count > 0) {
-      _pagesize = count;
-    }
+    super.top(count);
     return this;
   }
 
+  /// close parentheses
+  @override
   ExpenseFilterBuilder get endBlock {
-    if (_addedBlocks.needEndBlock[_blockIndex]) {
-      parameters[parameters.length - 1].whereString += ' ) ';
-    }
-    _addedBlocks.needEndBlock.removeAt(_blockIndex);
-    _addedBlocks.waitingStartBlock.removeAt(_blockIndex);
-    _blockIndex--;
+    super.endBlock;
     return this;
   }
 
-  ExpenseFilterBuilder orderBy(var argFields) {
-    if (argFields != null) {
-      if (argFields is String) {
-        orderByList.add(argFields);
-      } else {
-        for (String s in argFields) {
-          if (s != null && s != '') orderByList.add(' $s ');
-        }
-      }
-    }
+  /// argFields might be String or List<String>.
+  /// Example 1: argFields='name, date'
+  /// Example 2: argFields = ['name', 'date']
+  @override
+  ExpenseFilterBuilder orderBy(dynamic argFields) {
+    super.orderBy(argFields);
     return this;
   }
 
-  ExpenseFilterBuilder orderByDesc(var argFields) {
-    if (argFields != null) {
-      if (argFields is String) {
-        orderByList.add('$argFields desc ');
-      } else {
-        for (String s in argFields) {
-          if (s != null && s != '') orderByList.add(' $s desc ');
-        }
-      }
-    }
+  /// argFields might be String or List<String>.
+  /// Example 1: argFields='field1, field2'
+  /// Example 2: argFields = ['field1', 'field2']
+  @override
+  ExpenseFilterBuilder orderByDesc(dynamic argFields) {
+    super.orderByDesc(argFields);
     return this;
   }
 
-  ExpenseFilterBuilder groupBy(var argFields) {
-    if (argFields != null) {
-      if (argFields is String) {
-        groupByList.add(' $argFields ');
-      } else {
-        for (String s in argFields) {
-          if (s != null && s != '') groupByList.add(' $s ');
-        }
-      }
-    }
+  /// argFields might be String or List<String>.
+  /// Example 1: argFields='field1, field2'
+  /// Example 2: argFields = ['field1', 'field2']
+  @override
+  ExpenseFilterBuilder groupBy(dynamic argFields) {
+    super.groupBy(argFields);
     return this;
   }
 
-  ExpenseField setField(ExpenseField field, String colName, DbType dbtype) {
+  /// argFields might be String or List<String>.
+  /// Example 1: argFields='name, date'
+  /// Example 2: argFields = ['name', 'date']
+  @override
+  ExpenseFilterBuilder having(dynamic argFields) {
+    super.having(argFields);
+    return this;
+  }
+
+  ExpenseField _setField(ExpenseField? field, String colName, DbType dbtype) {
     return ExpenseField(this)
       ..param = DbParameter(
-          dbType: dbtype,
-          columnName: colName,
-          wStartBlock: _addedBlocks.waitingStartBlock[_blockIndex]);
+          dbType: dbtype, columnName: colName, wStartBlock: openedBlock);
   }
 
-  ExpenseField _id;
+  ExpenseField? _id;
   ExpenseField get id {
-    return _id = setField(_id, 'id', DbType.integer);
+    return _id = _setField(_id, 'id', DbType.integer);
   }
 
-  ExpenseField _amount;
+  ExpenseField? _amount;
   ExpenseField get amount {
-    return _amount = setField(_amount, 'amount', DbType.real);
+    return _amount = _setField(_amount, 'amount', DbType.real);
   }
 
-  ExpenseField _description;
+  ExpenseField? _description;
   ExpenseField get description {
-    return _description = setField(_description, 'description', DbType.text);
+    return _description = _setField(_description, 'description', DbType.text);
   }
 
-  ExpenseField _total;
+  ExpenseField? _total;
   ExpenseField get total {
-    return _total = setField(_total, 'total', DbType.real);
+    return _total = _setField(_total, 'total', DbType.real);
   }
 
-  ExpenseField _isDeleted;
+  ExpenseField? _isDeleted;
   ExpenseField get isDeleted {
-    return _isDeleted = setField(_isDeleted, 'isDeleted', DbType.bool);
+    return _isDeleted = _setField(_isDeleted, 'isDeleted', DbType.bool);
   }
 
-  bool _getIsDeleted;
-
-  void _buildParameters() {
-    if (_page > 0 && _pagesize > 0) {
-      qparams
-        ..limit = _pagesize
-        ..offset = (_page - 1) * _pagesize;
-    } else {
-      qparams
-        ..limit = _pagesize
-        ..offset = _page;
-    }
-    for (DbParameter param in parameters) {
-      if (param.columnName != null) {
-        if (param.value is List) {
-          param.value = param.value
-              .toString()
-              .replaceAll('[', '')
-              .replaceAll(']', '')
-              .toString();
-          whereString += param.whereString
-              .replaceAll('{field}', param.columnName)
-              .replaceAll('?', param.value.toString());
-          param.value = null;
-        } else {
-          whereString +=
-              param.whereString.replaceAll('{field}', param.columnName);
-        }
-        switch (param.dbType) {
-          case DbType.bool:
-            param.value =
-                param.value == null ? null : param.value == true ? 1 : 0;
-            break;
-          case DbType.date:
-          case DbType.datetime:
-            param.value = param.value == null
-                ? null
-                : (param.value as DateTime).millisecondsSinceEpoch;
-            break;
-          default:
-        }
-
-        if (param.value != null) {
-          whereArguments.add(param.value);
-        }
-        if (param.value2 != null) {
-          whereArguments.add(param.value2);
-        }
-      } else {
-        whereString += param.whereString;
-      }
-    }
-    if (Expense._softDeleteActivated) {
-      if (whereString != '') {
-        whereString =
-            '${!_getIsDeleted ? 'ifnull(isDeleted,0)=0 AND' : ''} ($whereString)';
-      } else if (!_getIsDeleted) {
-        whereString = 'ifnull(isDeleted,0)=0';
-      }
-    }
-
-    if (whereString != '') {
-      qparams.whereString = whereString;
-    }
-    qparams
-      ..whereArguments = whereArguments
-      ..groupBy = groupByList.join(',')
-      ..orderBy = orderByList.join(',');
-  }
-
-  /// <summary>
-  /// Deletes List<Expense> batch by query
-
-  /// <returns>BoolResult res.success=Deleted, not res.success=Can not deleted
+  /// Deletes List<Expense> bulk by query
+  ///
+  /// <returns>BoolResult res.success= true (Deleted), false (Could not be deleted)
+  @override
   Future<BoolResult> delete([bool hardDelete = false]) async {
-    _buildParameters();
-    var r = BoolResult();
-    if (Expense._softDeleteActivated && !hardDelete) {
-      r = await _obj._mnExpense.updateBatch(qparams, {'isDeleted': 1});
+    buildParameters();
+    var r = BoolResult(success: false);
+
+    if (_softDeleteActivated && !hardDelete) {
+      r = await _mnExpense!.updateBatch(qparams, {'isDeleted': 1});
     } else {
-      r = await _obj._mnExpense.delete(qparams);
+      r = await _mnExpense!.delete(qparams);
     }
     return r;
   }
 
+  /// Recover List<Expense> bulk by query
+  @override
   Future<BoolResult> recover() async {
-    _getIsDeleted = true;
-    _buildParameters();
-    print('SQFENTITIY: recover Expense batch invoked');
-    return _obj._mnExpense.updateBatch(qparams, {'isDeleted': 0});
+    buildParameters(getIsDeleted: true);
+    debugPrint('SQFENTITIY: recover Expense bulk invoked');
+    return _mnExpense!.updateBatch(qparams, {'isDeleted': 0});
   }
 
+  /// using:
+  /// update({'fieldName': Value})
+  /// fieldName must be String. Value is dynamic, it can be any of the (int, bool, String.. )
+  @override
   Future<BoolResult> update(Map<String, dynamic> values) {
-    _buildParameters();
-    return _obj._mnExpense.updateBatch(qparams, values);
+    buildParameters();
+    if (qparams.limit! > 0 || qparams.offset! > 0) {
+      qparams.whereString =
+          'id IN (SELECT id from expense ${qparams.whereString!.isNotEmpty ? 'WHERE ${qparams.whereString}' : ''}${qparams.limit! > 0 ? ' LIMIT ${qparams.limit}' : ''}${qparams.offset! > 0 ? ' OFFSET ${qparams.offset}' : ''})';
+    }
+    return _mnExpense!.updateBatch(qparams, values);
   }
 
-  /// This method always returns ExpenseObj if exist, otherwise returns null
-  /// <returns>List<Expense>
-  Future<Expense> toSingle([VoidCallback expense(Expense o)]) async {
-    _pagesize = 1;
-    _buildParameters();
-    final objFuture = _obj._mnExpense.toList(qparams);
+  /// This method always returns [Expense] Obj if exist, otherwise returns null
+  /// bool preload: if true, loads all related child objects (Set preload to true if you want to load all fields related to child or parent)
+  /// ex: toSingle(preload:true) -> Loads all related objects
+  /// List<String> preloadFields: specify the fields you want to preload (preload parameter's value should also be "true")
+  /// ex: toSingle(preload:true, preloadFields:['plField1','plField2'... etc])  -> Loads only certain fields what you specified
+  /// bool loadParents: if true, loads all parent objects until the object has no parent
+
+  /// <returns> Expense?
+  @override
+  Future<Expense?> toSingle(
+      {bool preload = false,
+      List<String>? preloadFields,
+      bool loadParents = false,
+      List<String>? loadedFields}) async {
+    buildParameters(pSize: 1);
+    final objFuture = _mnExpense!.toList(qparams);
     final data = await objFuture;
-    Expense retVal;
+    Expense? obj;
     if (data.isNotEmpty) {
-      retVal = Expense.fromMap(data[0] as Map<String, dynamic>);
+      obj = Expense.fromMap(data[0] as Map<String, dynamic>);
     } else {
-      retVal = null;
+      obj = null;
     }
-    if (expense != null) {
-      expense(retVal);
-    }
-    return retVal;
+    return obj;
   }
 
-  /// This method always returns int.
+  /// This method always returns [Expense]
+  /// bool preload: if true, loads all related child objects (Set preload to true if you want to load all fields related to child or parent)
+  /// ex: toSingle(preload:true) -> Loads all related objects
+  /// List<String> preloadFields: specify the fields you want to preload (preload parameter's value should also be "true")
+  /// ex: toSingle(preload:true, preloadFields:['plField1','plField2'... etc])  -> Loads only certain fields what you specified
+  /// bool loadParents: if true, loads all parent objects until the object has no parent
+
+  /// <returns> Expense?
+  @override
+  Future<Expense> toSingleOrDefault(
+      {bool preload = false,
+      List<String>? preloadFields,
+      bool loadParents = false,
+      List<String>? loadedFields}) async {
+    return await toSingle(
+            preload: preload,
+            preloadFields: preloadFields,
+            loadParents: loadParents,
+            loadedFields: loadedFields) ??
+        Expense();
+  }
+
+  /// This method returns int. [Expense]
   /// <returns>int
-  Future<int> toCount([VoidCallback expenseCount(int c)]) async {
-    _buildParameters();
+  @override
+  Future<int> toCount([VoidCallback Function(int c)? expenseCount]) async {
+    buildParameters();
     qparams.selectColumns = ['COUNT(1) AS CNT'];
-    final expensesFuture = await _obj._mnExpense.toList(qparams);
+    final expensesFuture = await _mnExpense!.toList(qparams);
     final int count = expensesFuture[0]['CNT'] as int;
     if (expenseCount != null) {
       expenseCount(count);
@@ -827,111 +760,84 @@ class ExpenseFilterBuilder extends SearchCriteria {
     return count;
   }
 
-  /// This method always returns List<Expense>.
+  /// This method returns List<Expense> [Expense]
+  /// bool preload: if true, loads all related child objects (Set preload to true if you want to load all fields related to child or parent)
+  /// ex: toList(preload:true) -> Loads all related objects
+  /// List<String> preloadFields: specify the fields you want to preload (preload parameter's value should also be "true")
+  /// ex: toList(preload:true, preloadFields:['plField1','plField2'... etc])  -> Loads only certain fields what you specified
+  /// bool loadParents: if true, loads all parent objects until the object has no parent
+
   /// <returns>List<Expense>
+  @override
   Future<List<Expense>> toList(
-      [VoidCallback expenseList(List<Expense> o)]) async {
-    final List<Expense> expensesData = List<Expense>();
+      {bool preload = false,
+      List<String>? preloadFields,
+      bool loadParents = false,
+      List<String>? loadedFields}) async {
     final data = await toMapList();
-    final int count = data.length;
-    for (int i = 0; i < count; i++) {
-      expensesData.add(Expense.fromMap(data[i] as Map<String, dynamic>));
-    }
-    if (expenseList != null) expenseList(expensesData);
+    final List<Expense> expensesData = await Expense.fromMapList(data,
+        preload: preload,
+        preloadFields: preloadFields,
+        loadParents: loadParents,
+        loadedFields: loadedFields,
+        setDefaultValues: qparams.selectColumns == null);
     return expensesData;
   }
 
-  /// This method always returns Json String
+  /// This method returns Json String [Expense]
+  @override
   Future<String> toJson() async {
-    final list = List<dynamic>();
+    final list = <dynamic>[];
     final data = await toList();
     for (var o in data) {
-      list.add(o.toMap());
+      list.add(o.toMap(forJson: true));
     }
     return json.encode(list);
   }
 
-  /// This method always returns Json String.
+  /// This method returns Json String. [Expense]
+  @override
   Future<String> toJsonWithChilds() async {
-    final list = List<dynamic>();
+    final list = <dynamic>[];
     final data = await toList();
     for (var o in data) {
-      list.add(await o.toMapWithChilds());
+      list.add(await o.toMapWithChildren(false, true));
     }
     return json.encode(list);
   }
 
-  /// This method always returns List<dynamic>.
+  /// This method returns List<dynamic>. [Expense]
   /// <returns>List<dynamic>
+  @override
   Future<List<dynamic>> toMapList() async {
-    _buildParameters();
-    return await _obj._mnExpense.toList(qparams);
+    buildParameters();
+    return await _mnExpense!.toList(qparams);
   }
 
-  /// Returns List<DropdownMenuItem<Expense>>
-  Future<List<DropdownMenuItem<Expense>>> toDropDownMenu(
-      String displayTextColumn,
-      [VoidCallback dropDownMenu(List<DropdownMenuItem<Expense>> o)]) async {
-    _buildParameters();
-    final expensesFuture = _obj._mnExpense.toList(qparams);
-
-    final data = await expensesFuture;
-    final int count = data.length;
-    final List<DropdownMenuItem<Expense>> items = List()
-      ..add(DropdownMenuItem(
-        value: Expense(),
-        child: Text('Select Expense'),
-      ));
-    for (int i = 0; i < count; i++) {
-      items.add(
-        DropdownMenuItem(
-          value: Expense.fromMap(data[i] as Map<String, dynamic>),
-          child: Text(data[i][displayTextColumn].toString()),
-        ),
-      );
+  /// This method returns Primary Key List SQL and Parameters retVal = Map<String,dynamic>. [Expense]
+  /// retVal['sql'] = SQL statement string, retVal['args'] = whereArguments List<dynamic>;
+  /// <returns>List<String>
+  @override
+  Map<String, dynamic> toListPrimaryKeySQL([bool buildParams = true]) {
+    final Map<String, dynamic> _retVal = <String, dynamic>{};
+    if (buildParams) {
+      buildParameters();
     }
-    if (dropDownMenu != null) {
-      dropDownMenu(items);
-    }
-    return items;
+    _retVal['sql'] = 'SELECT `id` FROM expense WHERE ${qparams.whereString}';
+    _retVal['args'] = qparams.whereArguments;
+    return _retVal;
   }
 
-  /// Returns List<DropdownMenuItem<int>>
-  Future<List<DropdownMenuItem<int>>> toDropDownMenuInt(
-      String displayTextColumn,
-      [VoidCallback dropDownMenu(List<DropdownMenuItem<int>> o)]) async {
-    _buildParameters();
-    qparams.selectColumns = ['id', displayTextColumn];
-    final expensesFuture = _obj._mnExpense.toList(qparams);
-
-    final data = await expensesFuture;
-    final int count = data.length;
-    final List<DropdownMenuItem<int>> items = List()
-      ..add(DropdownMenuItem(
-        value: 0,
-        child: Text('Select Expense'),
-      ));
-    for (int i = 0; i < count; i++) {
-      items.add(
-        DropdownMenuItem(
-          value: data[i]['id'] as int,
-          child: Text(data[i][displayTextColumn].toString()),
-        ),
-      );
-    }
-    if (dropDownMenu != null) {
-      dropDownMenu(items);
-    }
-    return items;
-  }
-
-  /// This method always returns Primary Key List<int>.
+  /// This method returns Primary Key List<int>.
   /// <returns>List<int>
-  Future<List<int>> toListPrimaryKey([bool buildParameters = true]) async {
-    if (buildParameters) _buildParameters();
-    final List<int> idData = List<int>();
+  @override
+  Future<List<int>> toListPrimaryKey([bool buildParams = true]) async {
+    if (buildParams) {
+      buildParameters();
+    }
+    final List<int> idData = <int>[];
     qparams.selectColumns = ['id'];
-    final idFuture = await _obj._mnExpense.toList(qparams);
+    final idFuture = await _mnExpense!.toList(qparams);
 
     final int count = idFuture.length;
     for (int i = 0; i < count; i++) {
@@ -940,39 +846,37 @@ class ExpenseFilterBuilder extends SearchCriteria {
     return idData;
   }
 
-  /// Returns List<dynamic> for selected columns. Use this method for 'groupBy' with min,max,avg..
+  /// Returns List<dynamic> for selected columns. Use this method for 'groupBy' with min,max,avg..  [Expense]
   /// Sample usage: (see EXAMPLE 4.2 at https://github.com/hhtokpinar/sqfEntity#group-by)
-  Future<List<dynamic>> toListObject(
-      [VoidCallback listObject(List<dynamic> o)]) async {
-    _buildParameters();
+  @override
+  Future<List<dynamic>> toListObject() async {
+    buildParameters();
 
-    final objectFuture = _obj._mnExpense.toList(qparams);
+    final objectFuture = _mnExpense!.toList(qparams);
 
-    final List<dynamic> objectsData = List<dynamic>();
+    final List<dynamic> objectsData = <dynamic>[];
     final data = await objectFuture;
     final int count = data.length;
     for (int i = 0; i < count; i++) {
       objectsData.add(data[i]);
-    }
-    if (listObject != null) {
-      listObject(objectsData);
     }
     return objectsData;
   }
 
   /// Returns List<String> for selected first column
   /// Sample usage: await Expense.select(columnsToSelect: ['columnName']).toListString()
+  @override
   Future<List<String>> toListString(
-      [VoidCallback listString(List<String> o)]) async {
-    _buildParameters();
+      [VoidCallback Function(List<String> o)? listString]) async {
+    buildParameters();
 
-    final objectFuture = _obj._mnExpense.toList(qparams);
+    final objectFuture = _mnExpense!.toList(qparams);
 
-    final List<String> objectsData = List<String>();
+    final List<String> objectsData = <String>[];
     final data = await objectFuture;
     final int count = data.length;
     for (int i = 0; i < count; i++) {
-      objectsData.add(data[i][qparams.selectColumns[0]].toString());
+      objectsData.add(data[i][qparams.selectColumns![0]].toString());
     }
     if (listString != null) {
       listString(objectsData);
@@ -984,30 +888,30 @@ class ExpenseFilterBuilder extends SearchCriteria {
 
 // region ExpenseFields
 class ExpenseFields {
-  static TableField _fId;
+  static TableField? _fId;
   static TableField get id {
     return _fId = _fId ?? SqlSyntax.setField(_fId, 'id', DbType.integer);
   }
 
-  static TableField _fAmount;
+  static TableField? _fAmount;
   static TableField get amount {
     return _fAmount =
         _fAmount ?? SqlSyntax.setField(_fAmount, 'amount', DbType.real);
   }
 
-  static TableField _fDescription;
+  static TableField? _fDescription;
   static TableField get description {
     return _fDescription = _fDescription ??
         SqlSyntax.setField(_fDescription, 'description', DbType.text);
   }
 
-  static TableField _fTotal;
+  static TableField? _fTotal;
   static TableField get total {
     return _fTotal =
         _fTotal ?? SqlSyntax.setField(_fTotal, 'total', DbType.real);
   }
 
-  static TableField _fIsDeleted;
+  static TableField? _fIsDeleted;
   static TableField get isDeleted {
     return _fIsDeleted = _fIsDeleted ??
         SqlSyntax.setField(_fIsDeleted, 'isDeleted', DbType.integer);
@@ -1018,9 +922,13 @@ class ExpenseFields {
 //region ExpenseManager
 class ExpenseManager extends SqfEntityProvider {
   ExpenseManager()
-      : super(ExpenseModal(), tableName: _tableName, colId: _colId);
-  static String _tableName = 'expense';
-  static String _colId = 'id';
+      : super(ExpenseModal(),
+            tableName: _tableName,
+            primaryKeyList: _primaryKeyList,
+            whereStr: _whereStr);
+  static const String _tableName = 'expense';
+  static const List<String> _primaryKeyList = ['id'];
+  static const String _whereStr = 'id=?';
 }
 
 //endregion ExpenseManager
@@ -1028,7 +936,7 @@ class ExpenseManager extends SqfEntityProvider {
 class IdentitySequence {
   /// Assigns a new value when it is triggered and returns the new value
   /// returns Future<int>
-  Future<int> nextVal([VoidCallback nextval(int o)]) async {
+  Future<int> nextVal([VoidCallback Function(int o)? nextval]) async {
     final val = await ExpenseModalSequenceManager()
         .sequence(SequenceIdentitySequence.getInstance, true);
     if (nextval != null) {
@@ -1039,7 +947,7 @@ class IdentitySequence {
 
   /// Get the current value
   /// returns Future<int>
-  Future<int> currentVal([VoidCallback currentval(int o)]) async {
+  Future<int> currentVal([VoidCallback Function(int o)? currentval]) async {
     final val = await ExpenseModalSequenceManager()
         .sequence(SequenceIdentitySequence.getInstance, false);
     if (currentval != null) {
@@ -1050,7 +958,7 @@ class IdentitySequence {
 
   /// Reset sequence to start value
   /// returns start value
-  Future<int> reset([VoidCallback currentval(int o)]) async {
+  Future<int> reset([VoidCallback Function(int o)? currentval]) async {
     final val = await ExpenseModalSequenceManager()
         .sequence(SequenceIdentitySequence.getInstance, false, reset: true);
     if (currentval != null) {
