@@ -1,9 +1,9 @@
- import 'package:flutter/material.dart';
+import 'package:flutter/material.dart';
 
 final Color buttoncolor = Colors.red;
-final Color bottomSheetBackgroundColor = Colors.grey[850];
+final Color bottomSheetBackgroundColor = Colors.grey.shade800;
 final Color floatingIconColor = Colors.black;
 final Color floatingBackgroundColor = Colors.white;
-final Color scaffoldBackgroundColor = Colors.grey[900]; 
+final Color scaffoldBackgroundColor = Colors.grey.shade900;
 final Color dismissedRight = Colors.red;
-final Color dismissedLeft = Colors.grey[700];
+final Color dismissedLeft = Colors.grey.shade700;
